[Update] WebApiFactory
1.修正Controller參數轉換 : Put / Patch
2.新增Response物件 : ContactInfoRP
3.調整Controller/Command回傳 : DB物件 -> Response物件
4.調整Controller/Command/Service/Model名稱

+++ API/WebApiFactory/WebApiFactory/Models/Data/ContactInfo.cs
@@ -28,7 +28,7 @@
         /// <summary>
         /// 年齡
         /// </summary>
-        public int? Age { get; set; }
+        public short? Age { get; set; }
 
         /// <summary>
         /// 電話號碼
